style: rename `LanSessionRunner` to `GgrsSessionRunner` and move it to `networking` module.
This prepares for when we make the `GgrsSessionRunner` agnostic to LAN
or online games with a new network socket trait.
